pyproject.toml
[project]
name = "food-recommendation-system-service-kge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.8",
    "networkx>=3.4.2",
    "pandas>=2.2.3",
    "pykeen>=1.11.0",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
]
